Implement new SR calculator
This commit uses the "Koohii" library by Francesco149, a pure Java
implementation of the pp-calculator "oppai-ng", coming from the same
author.

This allows for a dynamically recomputable star rating when mods are
applied/removed. Additionally, this allows for future local PP
computation, which may be of use to some players.