Add option to lock cursor to window
Locking the cursor to the window prevents the mouse from leaving the
window. This is useful in windowed mode where accidentally leaving the
window results in the game being paused.

The F9 key has been globally bound to toggle this option. osu! uses F9
to toggle its extended chat window, but opsu! is unlikely to implement a
chat feature. As well, it is close to the F10 key, which toggles mouse
buttons.